// src/lib/types.ts
export interface HtmlInputrops extends StrictHtmlInputrops {
  [key: string]: any;
}

export interface StrictHtmlInputrops {
  type?: string;
}

export interface HtmlSpanProps extends StrictHtmlSpanProps {
  [key: string]: any;
}

export interface StrictHtmlSpanProps {
  children?: React.ReactNode;
}

// ======================================================
// Types
// ======================================================

export type SemanticShorthandItemFunc<TProps> = (component: React.ReactType<TProps>, props: TProps, children?: React.ReactNode | React.ReactNodeArray) => React.ReactElement<any> | null;
export type SemanticShorthandContent = React.ReactNode;
export type SemanticShorthandItem<TProps> = React.ReactNode | TProps | SemanticShorthandItemFunc<TProps>;
export type SemanticShorthandCollection<TProps> = SemanticShorthandItem<TProps>[];

// ======================================================
// Styling
// ======================================================

export type SemanticCOLORS = 'red' | 'orange' | 'yellow' | 'olive' | 'green' | 'teal' | 'blue' | 'violet' | 'purple' | 'pink' | 'brown' | 'grey' | 'black';
export type SemanticSIZES = 'mini' | 'tiny' | 'small' | 'medium' | 'large' | 'big' | 'huge' | 'massive';
// ...
export interface FCX<P> extends React.FC<P> {

  create: Function;
}